+++ packages/article-skeleton/showcase-helper.js
@@ -1,4 +1,4 @@
-/* eslint-disable react/prop-types */
+/* eslint-disable react/prop-types,no-shadow */
 import React from "react";
 import { Text, View } from "react-native";
 import pick from "lodash.pick";
@@ -114,6 +114,20 @@ const renderArticleSkeleton = ({
   const data = fullArticleFixture(config);
   const showHeader = header ? () => <TestHeader /> : () => null;
 
+  const olympicsKeys = {
+    staging: {
+      endpoint: "https://olympics-embed-staging.pamedia.io",
+      authToken: "6i3DuEwbVhr2Fht6",
+      gamesCode: "OG2020-TR2"
+    },
+    prod: {
+      endpoint: "https://olympics-embed.pamedia.io",
+      authToken: "6i3DuEwbVhr2Fht6",
+      gamesCode: "OG2020-TR2"
+    }
+  };
+  const endpoint = select("OlympicsKeys", olympicsKeys, olympicsKeys.staging);
+
   return (
     <MockBookmarksProvider otherMocks={mocks} delay={1000} articleId={data.id}>
       <ContextProviderWithDefaults value={{ theme: { scale, sectionColour } }}>
@@ -142,6 +156,7 @@ const renderArticleSkeleton = ({
           )}
           onVideoPress={preventDefaultedAction(decorateAction)("onVideoPress")}
           onViewableItemsChanged={() => null}
+          olympicsKeys={endpoint}
         />
       </ContextProviderWithDefaults>
     </MockBookmarksProvider>