+++ packages/article-skeleton/src/article-skeleton.web.js
@@ -52,7 +52,8 @@ const ArticleSkeleton = ({
   additionalRelatedArticlesFlag,
   algoliaSearchKeys,
   latestFromSectionFlag,
-  latestFromSection
+  latestFromSection,
+  olympicsKeys
 }) => {
   const {
     commentsEnabled,
@@ -192,6 +193,7 @@ const ArticleSkeleton = ({
                       paidContentClassName={paidContentClassName}
                       template={template}
                       isPreview={isPreview}
+                      olympicsKeys={olympicsKeys}
                     />
                   )}
                   <PaywallPortal