+++ packages/ts-components/src/index.ts
@@ -39,3 +39,11 @@ export {
 export {
   LatestFromSection
 } from './components/latest-from-section/LatestFromSection';
+
+export {
+  OlympicsMedalTable
+} from './components/olympics/medal-table/OlympicsMedalTable';
+
+export {
+  OlympicsSchedule
+} from './components/olympics/schedule/OlympicsSchedule';