+++ packages/article-magazine-comment/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.29.16](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-comment@3.29.15...@times-components/article-magazine-comment@3.29.16) (2021-07-26)
+
+**Note:** Version bump only for package @times-components/article-magazine-comment
+
+
+
+
+
+## [3.29.15](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-comment@3.29.14...@times-components/article-magazine-comment@3.29.15) (2021-07-26)
+
+**Note:** Version bump only for package @times-components/article-magazine-comment
+
+
+
+
+
+## [3.29.14](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-comment@3.29.13...@times-components/article-magazine-comment@3.29.14) (2021-07-23)
+
+**Note:** Version bump only for package @times-components/article-magazine-comment
+
+
+
+
+
+## [3.29.13](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-comment@3.29.12...@times-components/article-magazine-comment@3.29.13) (2021-07-22)
+
+**Note:** Version bump only for package @times-components/article-magazine-comment
+
+
+
+
+
+## [3.29.12](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-comment@3.29.11...@times-components/article-magazine-comment@3.29.12) (2021-07-21)
+
+**Note:** Version bump only for package @times-components/article-magazine-comment
+
+
+
+
+
 ## [3.29.11](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-comment@3.29.10...@times-components/article-magazine-comment@3.29.11) (2021-07-20)
 
 **Note:** Version bump only for package @times-components/article-magazine-comment